+++ src/lib/components/ReservationForm.svelte
@@ -9,9 +9,8 @@
 	import { users, reservations, buoys } from '$lib/stores';
 	import { beforeResCutoff } from '$lib/reservationTimes.js';
 	import { Settings } from '$lib/settings';
-	import { checkSpaceAvailable } from '$lib/validationUtils.js';
 
-	import { validateBuddies, checkNoOverlappingRsvs } from '$utils/validation';
+	import { validateBuddies, checkSpaceAvailable } from '$utils/validation';
 
 	import {
 		augmentRsv,
@@ -68,7 +67,7 @@
 			return;
 		}
 
-		result = checkSpaceAvailable(Settings, $buoys, submitted, $reservations);
+		result = checkSpaceAvailable($buoys, submitted, $reservations);
 		if (result.status === 'error') {
 			popup(result.message);
 			cancel();
