+++ src/types.ts
@@ -1,5 +1,6 @@
-import type { Reservations } from './lib/server/xata.codegen';
+import type { Reservations, Buoys } from './lib/server/xata.codegen';
 
+// TODO: fix this annoying type shape rsv?.user?.id != sub.user
 export type ReservationData = Omit<Reservations, 'category' | 'status'> & {
 	category: ReservationCategory;
 	status: ReservationStatus;
@@ -27,8 +28,11 @@ export enum ReservationStatus {
 
 export type ReservationPeriod = 'upcoming' | 'past';
 
+// TODO: seems to be unrelated to Settings type from xata
 export type Setting = {
 	// TODO: fix type remove any
 	default: any;
 	entries: any[];
 };
+
+export type Buoy = Required<Buoys>;
